각종 버그 수정 및 구조 개선
- EnemyPlaneMedium4의 총알이 올바르지 않은 좌표에서 생성되던 문제 수정
- ShipCarrier_Aircraft의 적재 유닛 방향 반대로 수정
- 스테이지1 중간 보스 패턴이 정상적으로 작동하지 않던 문제 수정
- 스테이지4 타임라인 소폭 수정
- 스테이지4 중간 보스가 등장하지 않던 문제 수정
- 플레이어 부활 좌표 구조 개선
- 플레이어 좌표로 Vector2Int 타입을 쓰지 않도록 변경
- 플레이어 레이저 렌더링이 화면 상단을 조금 넘어갈 수 있도록 변경

+++ Assets/Scripts/SO Scripts/PlayerSpeedDatas.cs
@@ -9,8 +9,8 @@ public class PlayerSpeedDatas : ScriptableObject
     [Serializable]
     public struct PlayerSpeed
     {
-        public int defaultSpeed;
-        public int slowSpeed;
+        public float defaultSpeed;
+        public float slowSpeed;
     }
 
     public PlayerSpeed[] playerSpeed;
